wearlytics_backend/aiapp/serializers.py
from rest_framework import serializers
from .models import Chat

class ChatSerializer(serializers.ModelSerializer):
    class Meta:
        model = Chat
        fields = ['user_id', 'query', 'response', 'created_at']
