wearlytics_backend/requirements.txt
Django
djangorestframework
openai
google-generativeai
SpeechRecognition 
pydub